refactor(#230): move extra component youtube to its slide
