+++ webcomponents/slides/youtube/src/components/youtube/readme.md
@@ -1,0 +1,76 @@
+# deckgo-youtube
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Properties
+
+| Property     | Attribute     | Description | Type     | Default     |
+| ------------ | ------------- | ----------- | -------- | ----------- |
+| `frameTitle` | `frame-title` |             | `string` | `undefined` |
+| `height`     | `height`      |             | `number` | `undefined` |
+| `src`        | `src`         |             | `string` | `undefined` |
+| `width`      | `width`       |             | `number` | `undefined` |
+
+
+## Methods
+
+### `lazyLoadContent() => Promise<void>`
+
+
+
+#### Returns
+
+Type: `Promise<void>`
+
+
+
+### `pause() => Promise<void>`
+
+
+
+#### Returns
+
+Type: `Promise<void>`
+
+
+
+### `play() => Promise<void>`
+
+
+
+#### Returns
+
+Type: `Promise<void>`
+
+
+
+### `updateIFrame(width: number, height: number) => Promise<void>`
+
+
+
+#### Returns
+
+Type: `Promise<void>`
+
+
+
+
+## Dependencies
+
+### Used by
+
+ - [deckgo-slide-youtube](../slide)
+
+### Graph
+```mermaid
+graph TD;
+  deckgo-slide-youtube --> deckgo-youtube
+  style deckgo-youtube fill:#f9f,stroke:#333,stroke-width:4px
+```
+
+----------------------------------------------
+
+*Built with [StencilJS](https://stenciljs.com/)*
